+++ BasicShapes.h
@@ -15,9 +15,9 @@
 
 class Rect:public shape
 {
-	int hght, wdth;	//height and width of the recangle
+	double hght, wdth;	//height and width of the recangle
 public:
-	Rect(game* r_pGame, point ref, int r_hght, int r_wdth);
+	Rect(game* r_pGame, point ref, double r_hght, double r_wdth);
 	virtual void draw() const;
 	virtual void rotate();
 	virtual void resizeup();
@@ -33,9 +33,9 @@ public:
 class circle :public shape
 {
 	//Add data memebrs for class circle
-	int rad;
+	double rad;
 public:	
-	circle(game* r_pGame, point ref, int r);	//add more parameters for the constructor if needed
+	circle(game* r_pGame, point ref, double r);	//add more parameters for the constructor if needed
 	virtual void draw() const;
 	virtual void rotate();
 	virtual void resizeup();
@@ -46,10 +46,10 @@ public:
 
 class EquiTri :public shape
 {
-	int sidelen;
+	double sidelen;
 	double angle = 3.14/2;	//side length of the triangle
 public:
-	EquiTri(game* r_pGame, point ref, int r_sidelen);
+	EquiTri(game* r_pGame, point ref, double r_sidelen);
 	virtual void draw() const;
 	virtual void rotate();
 	virtual void resizeup();
